[improve][client] PIP-478: move the v4->v5 auth bridge into pulsar-client
Preparation for making the v5 authentication model the client's native driver. The bridge that
adapts a legacy v4 plugin into the v5 SPI lives in pulsar-client-v5, which depends on
pulsar-client — so the v4 client cannot reach it. Move it to where the v4 client can:

  pulsar-client-v5  org.apache.pulsar.client.impl.v5.auth.LegacyV4AuthenticationAdapter
  pulsar-client-v5  org.apache.pulsar.client.impl.v5.auth.TlsAuthentication
  pulsar-client-v5  org.apache.pulsar.client.impl.v5.V5AuthenticationLoader
    ->  pulsar-client  org.apache.pulsar.client.impl.auth.v5.*

V5AuthenticationLoader and its create(...) overloads widen to public, since callers are now in
another module. No behaviour changes.

Two adjustments the move forces:

- LegacyV4AuthenticationAdapter logged through slf4j, which is not on pulsar-client's main
  compile classpath; it now uses slog (@CustomLog) like the rest of the module.
- The relocated tests used SimpleAuthInitContext / SimpleAuthCallContext, which are
  package-private in pulsar-client-v5. They now build the same contexts through the public
  V5AuthContexts factory. V5AuthenticationLoaderTest stays in pulsar-client-v5 because it
  exercises V5ToV4AuthenticationAdapter, which does not move.
